consolidate aave demo magic into demoMagic
