+++ packages/demo/frontend/src/components/earn/frontendWalletOperations.ts
@@ -25,7 +25,7 @@ import {
   MorphoUSDCBorrowOPDemo,
 } from '@/constants/markets'
 import { mintDemoAsset } from '@/utils/demoAssetMinting'
-import { mintMirrorUsdcDemo, removeMirrorUsdcDemo } from '@/utils/borrowMirror'
+import { mirrorBorrowReceipt } from '@/utils/demoMagic'
 
 export type FrontendWalletOperationsWallet = Pick<
   Wallet,
@@ -137,18 +137,6 @@ export function buildFrontendBorrowOperations(
       | StubCollateralParams
       | StubRepayParams,
   ) => buildWalletBorrowParams(params, wallet.address)
-  // Mirror a real Aave borrow/repay as USDC_DEMO (silent, best-effort). The
-  // server-wallet path mirrors in the backend; the in-browser wallet does it
-  // here. No-op for Morpho or zero-amount receipts.
-  const mirror = (
-    marketId: BorrowQuoteParams['marketId'],
-    receipt: { borrowAmount: bigint },
-    settle: typeof mintMirrorUsdcDemo,
-  ) => {
-    if (marketId.kind === 'aave-v3' && receipt.borrowAmount > 0n) {
-      void settle(wallet, receipt.borrowAmount)
-    }
-  }
   return {
     getTokenBalances: async () => wallet.getBalance(),
     getMarkets: async () => actions.borrow.getMarkets(),
@@ -162,12 +150,12 @@ export function buildFrontendBorrowOperations(
     getQuote: async (params) => actions.borrow.getQuote(withParams(params)),
     openPosition: async (_walletAddress, params) => {
       const receipt = await wallet.borrow.openPosition(withParams(params))
-      mirror(params.marketId, receipt, mintMirrorUsdcDemo)
+      mirrorBorrowReceipt(wallet, params.marketId, 'mint', receipt)
       return receipt
     },
     closePosition: async (_walletAddress, params) => {
       const receipt = await wallet.borrow.closePosition(withParams(params))
-      mirror(params.marketId, receipt, removeMirrorUsdcDemo)
+      mirrorBorrowReceipt(wallet, params.marketId, 'remove', receipt)
       return receipt
     },
     depositCollateral: async (_walletAddress, params) =>
@@ -176,7 +164,7 @@ export function buildFrontendBorrowOperations(
       wallet.borrow.withdrawCollateral(withParams(params)),
     repay: async (_walletAddress, params) => {
       const receipt = await wallet.borrow.repay(withParams(params))
-      mirror(params.marketId, receipt, removeMirrorUsdcDemo)
+      mirrorBorrowReceipt(wallet, params.marketId, 'remove', receipt)
       return receipt
     },
   }
